refactor(queue): extract throttling checks into telegram-solve-queue.throttling.lib.mjs
telegram-solve-queue.lib.mjs was 1374 lines, past the 1350-line early-warning
threshold of the CI file-headroom check (issue #1593).

checkSystemResources (RAM/CPU/disk) and checkApiLimits (Claude/Codex/GitHub)
are pure decision functions over cached limit readings; they only need the
queue's verbose flag, logger and throttle counter. They move to their own
module and take the queue explicitly, with thin delegating methods left on
SolveQueue so every caller and test keeps working. getLocale and
appendWaitingForCurrentCommand move with them.

telegram-solve-queue.lib.mjs: 1374 -> 1121 lines.

Refs #2175
