+++ src/telegram-solve-queue.lib.mjs
@@ -9,7 +9,7 @@
  *
  * @see https://github.com/link-assistant/hive-mind/issues/1041
  */
-import { getCachedClaudeLimits, getCachedCodexLimits, getCachedGitHubLimits, getCachedMemoryInfo, getCachedCpuInfo, getCachedDiskInfo, getLimitCache } from './limits.lib.mjs';
+import { getLimitCache } from './limits.lib.mjs';
 export { formatDuration, getRunningAgentProcesses, getRunningClaudeProcesses, getRunningCodexProcesses, getRunningGeminiProcesses, getRunningProcesses, getRunningQwenProcesses } from './telegram-solve-queue.helpers.lib.mjs';
 import { collectExecutingItems, formatDuration, formatQueueToolSection, formatWaitingReason, getRunningAgentProcesses, getRunningClaudeProcesses, getRunningCodexProcesses, getRunningGeminiProcesses, getRunningProcesses, getRunningQwenProcesses, getRunningSessionItems, groupQueueItemsByTool, reportDequeueDecision } from './telegram-solve-queue.helpers.lib.mjs';
 export { QUEUE_CONFIG, THRESHOLD_STRATEGIES } from './queue-config.lib.mjs';
@@ -20,6 +20,8 @@ import { canonicalizeGitHubUrl as canonicalizeQueueUrl } from './github-url-pars
 import { t } from './i18n.lib.mjs';
 import { safeEditMessageText } from './telegram-safe-reply.lib.mjs';
 import { lt } from './limits-i18n.lib.mjs';
+// Issue #2175: throttling decisions live in their own module to keep this file under the 1350-line warning threshold.
+import { checkApiLimits as checkApiLimitsImpl, checkSystemResources as checkSystemResourcesImpl, getLocale } from './telegram-solve-queue.throttling.lib.mjs';
 export const QueueItemStatus = {
   QUEUED: 'queued',
   WAITING: 'waiting',
@@ -28,13 +30,6 @@ export const QueueItemStatus = {
   FAILED: 'failed',
   CANCELLED: 'cancelled',
 };
-function getLocale(options = {}) {
-  if (typeof options === 'string') return options;
-  return options?.locale || null;
-}
-function appendWaitingForCurrentCommand(reason, locale) {
-  return `${reason} (${lt('queue_waiting_current_command', {}, { locale })})`;
-}
 function appendRemainingDuration(reason, ms, locale) {
   return `${reason} (${lt('remaining', { duration: formatDuration(ms, { locale }) }, { locale })})`;
 }
@@ -652,282 +647,34 @@ export class SolveQueue {
     };
   }
   /**
-   * Check system resources (RAM, CPU, disk) using cached values
+   * Check system resources (RAM, CPU, disk) using cached values.
    *
-   * Uses 5-minute load average for CPU instead of instantaneous usage.
-   * This provides a more stable metric that isn't affected by brief spikes
-   * during claude process startup.
+   * Issue #2175: the implementation lives in
+   * telegram-solve-queue.throttling.lib.mjs so this file stays under the
+   * 1350-line early-warning threshold (issue #1593).
    *
-   * Resource threshold modes are now configurable via HIVE_MIND_QUEUE_CONFIG:
-   * - 'reject': Immediately reject the command, no queueing
-   * - 'enqueue': Block all commands unconditionally until metric drops
-   * - 'dequeue-one-at-a-time': Allow one command when above threshold
-   *
-   * Default strategies:
-   * - RAM: enqueue
-   * - CPU: enqueue
-   * - DISK: enqueue (waits until disk drops below the threshold)
-   *
-   * See: https://github.com/link-assistant/hive-mind/issues/1155
-   * See: https://github.com/link-assistant/hive-mind/issues/1253
-   * See: https://github.com/link-assistant/hive-mind/issues/1981
-   *
-   * @param {number} totalProcessing - Total processing count (queue + external claude processes)
+   * @param {number} totalProcessing - Total processing count (queue + external tool processes)
+   * @param {object} [options]
    * @returns {Promise<{ok: boolean, reasons: string[], oneAtATime: boolean, rejected: boolean, rejectReason: string|null}>}
    */
   async checkSystemResources(totalProcessing = 0, options = {}) {
-    const locale = getLocale(options);
-    const reasons = [];
-    let oneAtATime = false;
-    let rejected = false;
-    let rejectReason = null;
-    // Check RAM (using cached value)
-    const memResult = await getCachedMemoryInfo(this.verbose);
-    if (memResult.success) {
-      const usedRatio = memResult.memory.usedPercentage / 100;
-      if (usedRatio >= QUEUE_CONFIG.thresholds.ram.value) {
-        const reason = formatWaitingReason('ram', memResult.memory.usedPercentage, QUEUE_CONFIG.thresholds.ram.value, { locale });
-        const strategy = QUEUE_CONFIG.thresholds.ram.strategy;
-        this.recordThrottle(`ram_${strategy}`);
-        if (strategy === 'reject') {
-          rejected = true;
-          rejectReason = reason;
-        } else if (strategy === 'dequeue-one-at-a-time') {
-          oneAtATime = true;
-          if (totalProcessing > 0) {
-            reasons.push(appendWaitingForCurrentCommand(reason, locale));
-          }
-        } else {
-          // 'enqueue' - block unconditionally
-          reasons.push(reason);
-        }
-      }
-    }
-    // Check CPU using 5-minute load average (more stable than 1-minute)
-    const cpuResult = await getCachedCpuInfo(this.verbose);
-    if (cpuResult.success) {
-      // Use loadAvg5 (5-minute average) instead of usagePercentage (1-minute based)
-      // This provides a more stable metric that isn't affected by transient spikes
-      const loadAvg5 = cpuResult.cpuLoad.loadAvg5;
-      const cpuCount = cpuResult.cpuLoad.cpuCount;
-      // Calculate usage ratio: loadAvg5 / cpuCount
-      // Load average of 1.0 per CPU = 100% utilization
-      const usageRatio = loadAvg5 / cpuCount;
-      const usagePercent = Math.min(100, Math.round(usageRatio * 100));
-      if (this.verbose) {
-        this.log(`CPU 5m load avg: ${loadAvg5.toFixed(2)}, cpus: ${cpuCount}, usage: ${usagePercent}%`);
-      }
-      if (usageRatio >= QUEUE_CONFIG.thresholds.cpu.value) {
-        const reason = formatWaitingReason('cpu', usagePercent, QUEUE_CONFIG.thresholds.cpu.value, { locale });
-        const strategy = QUEUE_CONFIG.thresholds.cpu.strategy;
-        this.recordThrottle(`cpu_${strategy}`);
-        if (strategy === 'reject') {
-          rejected = true;
-          rejectReason = reason;
-        } else if (strategy === 'dequeue-one-at-a-time') {
-          oneAtATime = true;
-          if (totalProcessing > 0) {
-            reasons.push(appendWaitingForCurrentCommand(reason, locale));
-          }
-        } else {
-          // 'enqueue' - block unconditionally
-          reasons.push(reason);
-        }
-      }
-    }
-    // Check disk space (using cached value)
-    // Default strategy changed to 'reject' because queue is lost on restart anyway
-    // See: https://github.com/link-assistant/hive-mind/issues/1253
-    const diskResult = await getCachedDiskInfo(this.verbose);
-    if (diskResult.success) {
-      // Calculate usage from free percentage
-      const usedPercent = 100 - diskResult.diskSpace.freePercentage;
-      const usedRatio = usedPercent / 100;
-      if (usedRatio >= QUEUE_CONFIG.thresholds.disk.value) {
-        const reason = formatWaitingReason('disk', usedPercent, QUEUE_CONFIG.thresholds.disk.value, { locale });
-        const strategy = QUEUE_CONFIG.thresholds.disk.strategy;
-        this.recordThrottle(`disk_${strategy}`);
-        if (strategy === 'reject') {
-          rejected = true;
-          rejectReason = reason;
-        } else if (strategy === 'dequeue-one-at-a-time') {
-          oneAtATime = true;
-          if (totalProcessing > 0) {
-            reasons.push(appendWaitingForCurrentCommand(reason, locale));
-          }
-        } else {
-          // 'enqueue' - block unconditionally
-          reasons.push(reason);
-        }
-      }
-    }
-    return { ok: reasons.length === 0 && !rejected, reasons, oneAtATime, rejected, rejectReason };
+    return checkSystemResourcesImpl(this, totalProcessing, options);
   }
   /**
-   * Check API limits (Claude, GitHub) using cached values
+   * Check API limits (Claude, Codex, GitHub) using cached values.
    *
-   * Logic per issue #1133:
-   * - CLAUDE_5_HOUR_SESSION_THRESHOLD and CLAUDE_WEEKLY_THRESHOLD use one-at-a-time mode:
-   *   when above threshold, allow exactly one command, block if claudeProcessing > 0
-   * - GitHub threshold blocks unconditionally when exceeded (ultimate restriction)
+   * Issue #2175: the implementation lives in
+   * telegram-solve-queue.throttling.lib.mjs so this file stays under the
+   * 1350-line early-warning threshold (issue #1593).
    *
-   * Logic per issue #1159:
-   * - When tool is 'agent', 'gemini', or 'qwen', skip Claude-specific limits entirely since these tools use
-   *   different rate limiting backends. Only system resources and GitHub limits apply.
-   * - For Claude limits, only count Claude-specific processing items, not agent/codex/gemini/qwen items.
-   *   This allows non-Claude tasks to run in parallel even when Claude limits are reached.
-   *
-   * Logic per issue #1253:
-   * - All thresholds now support configurable strategies (reject, enqueue, dequeue-one-at-a-time)
-   * - Configuration via HIVE_MIND_QUEUE_CONFIG or individual env vars
-   *
-   * @param {boolean} hasRunningToolProcess - Whether matching tool processes are running (from pgrep)
-   * @param {number} toolProcessingCount - Count of matching tool items being processed in queue
-   * @param {string} tool - The tool being used ('claude', 'agent', 'codex', 'gemini', 'qwen', etc.)
+   * @param {boolean} hasRunningToolProcess
+   * @param {number} toolProcessingCount
+   * @param {string} tool
+   * @param {object} [options]
    * @returns {Promise<{ok: boolean, reasons: string[], oneAtATime: boolean, rejected: boolean, rejectReason: string|null}>}
    */
   async checkApiLimits(hasRunningToolProcess = false, toolProcessingCount = 0, tool = 'claude', options = {}) {
-    const locale = getLocale(options);
-    const reasons = [];
-    let oneAtATime = false;
-    let rejected = false;
-    let rejectReason = null;
-    // Apply Claude-specific limits only when tool is 'claude'
-    // Other tools (like 'agent', 'gemini', and 'qwen') use different rate limiting backends and are not
-    // affected by Claude API limits (5-hour session, weekly limits)
-    // See: https://github.com/link-assistant/hive-mind/issues/1159
-    const applyClaudeLimits = tool === 'claude';
-    const applyCodexLimits = tool === 'codex';
-    const totalToolProcessing = toolProcessingCount + (hasRunningToolProcess ? 1 : 0);
-    // Check Claude limits (using cached value)
-    // Only applied when tool is 'claude'
-    if (applyClaudeLimits) {
-      const claudeResult = await getCachedClaudeLimits(this.verbose);
-      if (claudeResult.success) {
-        const sessionPercent = claudeResult.usage.currentSession.percentage;
-        const weeklyPercent = claudeResult.usage.allModels.percentage;
-        // Session limit (5-hour)
-        // Configurable strategy via HIVE_MIND_QUEUE_CONFIG or HIVE_MIND_CLAUDE_5_HOUR_SESSION_STRATEGY
-        // See: https://github.com/link-assistant/hive-mind/issues/1133, #1159, #1253
-        if (sessionPercent !== null) {
-          const sessionRatio = sessionPercent / 100;
-          if (sessionRatio >= QUEUE_CONFIG.thresholds.claude5Hour.value) {
-            const reason = formatWaitingReason('claude_5_hour_session', sessionPercent, QUEUE_CONFIG.thresholds.claude5Hour.value, { locale });
-            const strategy = QUEUE_CONFIG.thresholds.claude5Hour.strategy;
-            this.recordThrottle(sessionRatio >= 1.0 ? 'claude_5_hour_session_100' : `claude_5_hour_session_${strategy}`);
-            if (strategy === 'reject') {
-              rejected = true;
-              rejectReason = reason;
-            } else if (strategy === 'dequeue-one-at-a-time') {
-              oneAtATime = true;
-              if (totalToolProcessing > 0) {
-                reasons.push(appendWaitingForCurrentCommand(reason, locale));
-              }
-            } else {
-              // 'enqueue' - block unconditionally
-              reasons.push(reason);
-            }
-          }
-        }
-        // Weekly limit
-        // Configurable strategy via HIVE_MIND_QUEUE_CONFIG or HIVE_MIND_CLAUDE_WEEKLY_STRATEGY
-        // See: https://github.com/link-assistant/hive-mind/issues/1133, #1159, #1253
-        if (weeklyPercent !== null) {
-          const weeklyRatio = weeklyPercent / 100;
-          if (weeklyRatio >= QUEUE_CONFIG.thresholds.claudeWeekly.value) {
-            const reason = formatWaitingReason('claude_weekly', weeklyPercent, QUEUE_CONFIG.thresholds.claudeWeekly.value, { locale });
-            const strategy = QUEUE_CONFIG.thresholds.claudeWeekly.strategy;
-            this.recordThrottle(weeklyRatio >= 1.0 ? 'claude_weekly_100' : `claude_weekly_${strategy}`);
-            if (strategy === 'reject') {
-              rejected = true;
-              rejectReason = reason;
-            } else if (strategy === 'dequeue-one-at-a-time') {
-              oneAtATime = true;
-              if (totalToolProcessing > 0) {
-                reasons.push(appendWaitingForCurrentCommand(reason, locale));
-              }
-            } else {
-              // 'enqueue' - block unconditionally
-              reasons.push(reason);
-            }
-          }
-        }
-      }
-    } else if (applyCodexLimits) {
-      const codexResult = await getCachedCodexLimits(this.verbose);
-      if (codexResult.success) {
-        const sessionPercent = codexResult.usage.currentSession.percentage;
-        const weeklyPercent = codexResult.usage.allModels.percentage;
-        if (sessionPercent !== null) {
-          const sessionRatio = sessionPercent / 100;
-          if (sessionRatio >= QUEUE_CONFIG.thresholds.codex5Hour.value) {
-            const reason = formatWaitingReason('codex_5_hour_session', sessionPercent, QUEUE_CONFIG.thresholds.codex5Hour.value, { locale });
-            const strategy = QUEUE_CONFIG.thresholds.codex5Hour.strategy;
-            this.recordThrottle(sessionRatio >= 1.0 ? 'codex_5_hour_session_100' : `codex_5_hour_session_${strategy}`);
-            if (strategy === 'reject') {
-              rejected = true;
-              rejectReason = reason;
-            } else if (strategy === 'dequeue-one-at-a-time') {
-              oneAtATime = true;
-              if (totalToolProcessing > 0) {
-                reasons.push(appendWaitingForCurrentCommand(reason, locale));
-              }
-            } else {
-              reasons.push(reason);
-            }
-          }
-        }
-        if (weeklyPercent !== null) {
-          const weeklyRatio = weeklyPercent / 100;
-          if (weeklyRatio >= QUEUE_CONFIG.thresholds.codexWeekly.value) {
-            const reason = formatWaitingReason('codex_weekly', weeklyPercent, QUEUE_CONFIG.thresholds.codexWeekly.value, { locale });
-            const strategy = QUEUE_CONFIG.thresholds.codexWeekly.strategy;
-            this.recordThrottle(weeklyRatio >= 1.0 ? 'codex_weekly_100' : `codex_weekly_${strategy}`);
-            if (strategy === 'reject') {
-              rejected = true;
-              rejectReason = reason;
-            } else if (strategy === 'dequeue-one-at-a-time') {
-              oneAtATime = true;
-              if (totalToolProcessing > 0) {
-                reasons.push(appendWaitingForCurrentCommand(reason, locale));
-              }
-            } else {
-              reasons.push(reason);
-            }
-          }
-        }
-      }
-    } else if (this.verbose) {
-      this.log(`Claude limits not applied for --tool ${tool}`);
-    }
-    // Check GitHub limits when the active tool already has a running process.
-    // This keeps the queue behavior aligned with the existing one-at-a-time throttling model.
-    // Configurable strategy via HIVE_MIND_QUEUE_CONFIG or HIVE_MIND_GITHUB_API_STRATEGY
-    if (hasRunningToolProcess) {
-      const githubResult = await getCachedGitHubLimits(this.verbose);
-      if (githubResult.success) {
-        const usedPercent = githubResult.githubRateLimit.usedPercentage;
-        const usedRatio = usedPercent / 100;
-        if (usedRatio >= QUEUE_CONFIG.thresholds.githubApi.value) {
-          const reason = formatWaitingReason('github', usedPercent, QUEUE_CONFIG.thresholds.githubApi.value, { locale });
-          const strategy = QUEUE_CONFIG.thresholds.githubApi.strategy;
-          this.recordThrottle(usedRatio >= 1.0 ? 'github_100' : `github_${strategy}`);
-          if (strategy === 'reject') {
-            rejected = true;
-            rejectReason = reason;
-          } else if (strategy === 'dequeue-one-at-a-time') {
-            oneAtATime = true;
-            if (totalToolProcessing > 0) {
-              reasons.push(appendWaitingForCurrentCommand(reason, locale));
-            }
-          } else {
-            // 'enqueue' - block unconditionally
-            reasons.push(reason);
-          }
-        }
-      }
-    }
-    return { ok: reasons.length === 0 && !rejected, reasons, oneAtATime, rejected, rejectReason };
+    return checkApiLimitsImpl(this, hasRunningToolProcess, toolProcessingCount, tool, options);
   }
   /**
    * Record a throttle event for statistics
